refactored and split code into semantic components, also removed unnecessary boilerplate
