--- src/components/Hero.vue
@@ -1,374 +1,0 @@
-<template>
-  <div id="hero">
-    <nav>
-      <ul>
-        <svg viewBox="0 0 551.39 142.09" xmlns="http://www.w3.org/2000/svg" width="150px">
-          <g fill="#2d3887" transform="translate(-27.85 -319.19)">
-            <path
-              d="m239 328.39a15.33 15.33 0 0 1 -2.06 7.06q-13 25.13-26 50.26-17.93 34.86-35.8 69.73a15.47 15.47 0 0 1 -2.06 3.22 7.07 7.07 0 0 1 -10.62.2 15.38 15.38 0 0 1 -2.46-3.8l-55.28-107.66q-3.61-7-7.25-14.06a10.23 10.23 0 0 1 -.73-8.56 7.54 7.54 0 0 1 5.84-5.11 15.48 15.48 0 0 1 3.33-.33c22.16 0 44.31-.07 66.46 0 10.4 0 20.79.24 31.18.23 8.08 0 16.15-.2 24.23-.34a16.09 16.09 0 0 1 4.74.52 8.52 8.52 0 0 1 6.48 8.64z"
-            ></path>
-            <path
-              d="m418.7 390q0-29.76 0-59.5a43.67 43.67 0 0 1 .35-5.25 6.14 6.14 0 0 1 5.41-5.58 31.77 31.77 0 0 1 4.78-.38q21.58 0 43.18 0a28 28 0 0 1 4.76.41 6 6 0 0 1 5.15 4.72 14.47 14.47 0 0 1 .48 3.55v18.95c0 3.39.5 4 3.86 3.79 5.43-.37 10.86-.1 16.29-.21h1.2c2.72.06 3.22.55 3.22 3.25 0 7.84-.05 15.67-.05 23.51q0 24.94 0 49.9v1.44c-.11 1.73-.62 2.28-2.33 2.37-1.19.06-2.4 0-3.6 0-5.19 0-10.39 0-15.59 0-2.77 0-3 .28-3 3v18a18.86 18.86 0 0 1 -.43 4 5.84 5.84 0 0 1 -4.78 4.78 18.87 18.87 0 0 1 -4 .44q-22.79 0-45.58 0a16.69 16.69 0 0 1 -3.8-.4 6 6 0 0 1 -5.11-5.58 54.58 54.58 0 0 1 -.38-5.73q0-29.73-.03-59.48z"
-            ></path>
-            <path
-              d="m326.16 390.85h-60.48c-1.28 0-2.56 0-3.83-.13-3.58-.35-5.45-2.13-6.13-5.66a17.86 17.86 0 0 1 -.28-3.34q0-26.52 0-53a18.73 18.73 0 0 1 .39-3.81 6.16 6.16 0 0 1 5.31-5.15 20.3 20.3 0 0 1 3.82-.3h122.86a20.34 20.34 0 0 1 3.58.28 6.61 6.61 0 0 1 5.63 5.19 14.84 14.84 0 0 1 .41 3.81v53.26a13.47 13.47 0 0 1 -.32 3.34c-.81 3.22-2.66 5-5.93 5.39a32.53 32.53 0 0 1 -3.83.15z"
-            ></path>
-            <path
-              d="m515.42 390.23q0-30.11 0-60.22c0-1.36 0-2.73.08-4.08.26-3.78 2.21-5.79 5.93-6.33a33.26 33.26 0 0 1 4.78-.3q21.24 0 42.47 0a30.29 30.29 0 0 1 4.77.38 5.94 5.94 0 0 1 5.39 5.27 28 28 0 0 1 .38 4.3q0 61.18 0 122.37a22 22 0 0 1 -.46 4.52 5.65 5.65 0 0 1 -4.56 4.69 15.71 15.71 0 0 1 -3.56.45q-23.5 0-47 0c-.48 0-1 0-1.44 0-4.3-.34-6.51-2.61-6.69-6.92-.06-1.36-.06-2.72-.06-4.08q-.03-30.04-.03-60.05z"
-            ></path>
-            <path
-              d="m57.83 319.28c3.62 0 9.52.24 15.44.05 5.73-.19 9.73 2.67 12.23 7.67 3 5.85 6 11.66 9 17.49q14.91 29 29.86 57.94c1.73 3.32 3.22 6.57 3 10.42a15.51 15.51 0 0 1 -1.85 6.64c-6.57 12.06-12.68 24.37-18.93 36.61a10.06 10.06 0 0 1 -2 3c-3.81 3.53-9.42 2.51-11.9-2.22-2.7-5.17-5.34-10.37-8-15.56l-54.1-105.32a23.55 23.55 0 0 1 -2.46-5.69c-1.24-5.61 1.84-10.12 7.51-10.85a32.7 32.7 0 0 1 3.83-.15z"
-            ></path>
-            <path
-              d="m327.37 399.43h32.38c2.72.07 3.25.56 3.26 3.24 0 4.48-.05 9-.05 13.44q0 17.75.05 35.5a23.35 23.35 0 0 1 -.15 3.35 6.62 6.62 0 0 1 -5.17 5.91 15.76 15.76 0 0 1 -4 .41h-53a14.44 14.44 0 0 1 -3.34-.26c-3.58-.86-5.51-3-5.89-6.63a21.31 21.31 0 0 1 -.08-2.4v-48.46c0-3.68.43-4.1 4.14-4.1z"
-            ></path>
-          </g>
-        </svg>
-        <li>
-          <a href="#">About</a>
-        </li>
-        <li>
-          <a href="#">Prizes</a>
-        </li>
-        <li>
-          <a href="#">Schedule</a>
-        </li>
-        <li>
-          <a href="#">Sponsors</a>
-        </li>
-        <li>
-          <a href="#">FAQ</a>
-        </li>
-        <li class="nav-register-link">
-          <a href="#">Register</a>
-        </li>
-        <li
-          style="height: 40px; border: 1px solid var(--color-text); transition: border 0.6s ease-out;"
-        ></li>
-        <li>
-          <svg
-            v-if="theme === 'light'"
-            width="25px"
-            style="transform: translateY(3.5px); cursor: pointer;"
-            @click="toggleTheme"
-            viewBox="0 0 222.59 218.11"
-            xmlns="http://www.w3.org/2000/svg"
-          >
-            <g transform="translate(-195.2 -286.74)">
-              <path
-                d="m376.78 472.87c.41 10.77.42 10.74-9.93 8.59-5.68-1.17-11.41-2.15-17.08-3.39-2.91-.63-4.47.29-5.81 3-3.94 7.94-8.21 15.73-12.45 23.78-7.29-5.36-14.5-10.5-21.51-15.91-2.48-1.91-4.22-2.22-6.83-.12-6 4.83-12.41 9.16-18.43 14-2.76 2.2-3.79 1.52-5.19-1.34-3.27-6.72-7.07-13.18-10.28-19.92-1.48-3.11-3.19-4.13-6.54-3.4-7.13 1.57-14.39 2.58-21.46 4.36-4.19 1.06-5.52-.06-5.13-4.14.17-1.81.07-3.65.08-5.47.76-3.9.15-7.83.35-11.74 0-2.44-.06-4.9.05-7.31-.41 1.05.41 2.44-.7 3.38 0-4.17-2.25-5.7-6.2-6.53-6.95-1.45-13.69-3.83-20.62-5.39-3.31-.75-3.58-1.89-2.28-4.7 3.33-7.2 6.33-14.56 9.67-21.76 1.17-2.5.85-3.59-1.37-5.62-6.63-6.05-14.48-10.81-19.92-18.16 6.67-5.85 13.16-11.92 20.11-17.42 2.32-1.84 2.29-3 1.27-5.25-3.16-6.93-6-14-9.36-20.85-1.95-4-1-5.35 3.15-6.24 6.66-1.42 13.13-3.76 19.81-5.07 4-.77 5.49-2.78 5.74-6.41 1.17.93.14 2.57.75 3.42-.23-2.3 0-4.78-.1-7.25-.2-4 .41-7.92-.35-11.84 0-1-.13-2-.11-3 .12-8.09.1-7.95 8.08-6.08 6.43 1.5 13 2.58 19.45 3.91 2.44.51 4-.19 5.14-2.53 3.63-7.26 7.56-14.38 11.15-21.67 1.24-2.51 2.29-2.58 4.26-.87 6.4 5.55 13.33 10.38 20.2 15.32 1.94 1.4 2.61.92 3.91 0 6.7-5 13.47-9.92 20.08-15 2.37-1.84 3.48-2 4.91 1.09 2.86 6.13 6.25 12 9.4 18s6.16 6.91 12.65 4.95c6.15-1.85 12.57-2.8 18.88-4.1 1.59-.33 2.56.27 2.54 2.07v8c-.76 3.9-.15 7.83-.35 11.74v7c.36-.9-.46-2.29.65-3.23-.3 6 4.55 6.16 8.37 7.22 6.51 1.82 13.08 3.38 19.64 5 3.11.79 1.61 2.83 1 4.28-2.62 6.26-5.4 12.45-8.13 18.65s-3.25 6.85 1.7 10.84c6.15 5 11.75 10.56 18.13 15.23-5.42 7.39-13.3 12.13-20 18.19-2.31 2.09-2.42 3.21-1.28 5.65 3.22 6.91 6.11 14 9.32 20.89 1.56 3.36 1 4.94-2.85 5.79-6.63 1.47-13.07 3.75-19.71 5.11-3.95.81-6.25 2.35-6.23 6.53-1.17-.93-.14-2.57-.75-3.42.23 2.3.05 4.78.1 7.25.26 4.02-.37 8.02.41 11.89z"
-                fill="#f4c425"
-              ></path>
-              <path d="m236.22 318.13c1.74 3.87.42 7.95.82 11.91-.32.47-.64.45-1 0z" fill="#f3c001"></path>
-              <path
-                d="m376 330c.41-4-.92-8 .82-11.91q.06 6 .14 11.91c-.36.51-.68.49-.96 0z"
-                fill="#f3c001"
-              ></path>
-              <path
-                d="m236.08 330h1v8h-.71c-.13-1.39-.27-2.79-.41-4.18.04-1.27.04-2.52.12-3.82z"
-                fill="#f4c214"
-              ></path>
-              <path
-                d="m376 330h1c.05 1.26.11 2.51.16 3.76-.14 1.39-.28 2.79-.41 4.18h-.75z"
-                fill="#f4c214"
-              ></path>
-              <path
-                d="m237 461c-.41 4 .92 8-.82 11.91q-.06-6-.14-11.91c.36-.49.68-.51.96 0z"
-                fill="#f3c001"
-              ></path>
-              <path
-                d="m376.78 472.87c-1.78-3.87-.42-7.95-.78-11.87.32-.45.64-.47 1 0q-.15 5.92-.22 11.87z"
-                fill="#f3c001"
-              ></path>
-              <path
-                d="m237 461h-1c-.05-1.26-.11-2.51-.16-3.76.14-1.39.28-2.79.41-4.18h.75z"
-                fill="#f4c214"
-              ></path>
-              <path
-                d="m376.92 461h-1v-8h.71c.13 1.39.27 2.79.41 4.18-.04 1.27-.04 2.52-.12 3.82z"
-                fill="#f4c214"
-              ></path>
-              <path
-                d="m232.09 387a75.38 75.38 0 0 1 64.91-65.91 76 76 0 0 0 18.94 0 75.38 75.38 0 0 1 64.97 65.91 60.84 60.84 0 0 0 0 16.94 75.11 75.11 0 0 1 -65.91 65.97 60.84 60.84 0 0 0 -16.94 0 75.1 75.1 0 0 1 -65.97-65.91 60.84 60.84 0 0 0 0-17z"
-                fill="#f7a705"
-              ></path>
-              <g fill="#f7a500">
-                <path d="m316 321.09a39.15 39.15 0 0 1 -18.94 0z"></path>
-                <path d="m232.09 387a31.24 31.24 0 0 1 0 16.94z"></path>
-                <path d="m380.91 404a31.24 31.24 0 0 1 0-16.94z"></path>
-                <path d="m298 469.91a31.24 31.24 0 0 1 16.94 0z"></path>
-              </g>
-            </g>
-          </svg>
-          <svg
-            v-else
-            width="25px"
-            style="transform: translateY(3.5px); cursor: pointer;"
-            @click="toggleTheme"
-            viewBox="0 0 223.99 225"
-            xmlns="http://www.w3.org/2000/svg"
-          >
-            <path
-              d="m106 0h13l19.6 2.67a9.73 9.73 0 0 1 -4.24 2.55c-39.09 14.31-62.72 54.67-55 93.86 12.47 62.94 86 90.76 137.25 51.92 2.36-1.78 4.89-3.34 7.34-5v1c-.93 1.87-2 3.7-2.78 5.63-13.6 33.75-38.06 55.91-72.49 66.88-8.92 2.85-18.47 3.71-27.73 5.49h-12.95a22.42 22.42 0 0 0 -3.33-.8c-47.58-4.11-86.17-35-100-80.4-2.16-7.08-3.14-14.52-4.67-21.8v-19c.39-2.44.7-4.89 1.17-7.32 8.61-44.68 34.38-74.68 77.25-89.59 8.82-3.09 18.36-4.09 27.58-6.09z"
-              fill="#f3f3fb"
-              transform="translate(-.01)"
-            ></path>
-          </svg>
-        </li>
-      </ul>
-    </nav>
-    <div class="hero-content">
-      <svg viewBox="0 0 551.39 142.09" xmlns="http://www.w3.org/2000/svg">
-        <g transform="translate(-27.85 -319.19)">
-          <path
-            d="m239 328.39a15.33 15.33 0 0 1 -2.06 7.06q-13 25.13-26 50.26-17.93 34.86-35.8 69.73a15.47 15.47 0 0 1 -2.06 3.22 7.07 7.07 0 0 1 -10.62.2 15.38 15.38 0 0 1 -2.46-3.8l-55.28-107.66q-3.61-7-7.25-14.06a10.23 10.23 0 0 1 -.73-8.56 7.54 7.54 0 0 1 5.84-5.11 15.48 15.48 0 0 1 3.33-.33c22.16 0 44.31-.07 66.46 0 10.4 0 20.79.24 31.18.23 8.08 0 16.15-.2 24.23-.34a16.09 16.09 0 0 1 4.74.52 8.52 8.52 0 0 1 6.48 8.64z"
-            fill="#a82175"
-          ></path>
-          <path
-            d="m418.7 390q0-29.76 0-59.5a43.67 43.67 0 0 1 .35-5.25 6.14 6.14 0 0 1 5.41-5.58 31.77 31.77 0 0 1 4.78-.38q21.58 0 43.18 0a28 28 0 0 1 4.76.41 6 6 0 0 1 5.15 4.72 14.47 14.47 0 0 1 .48 3.55v18.95c0 3.39.5 4 3.86 3.79 5.43-.37 10.86-.1 16.29-.21h1.2c2.72.06 3.22.55 3.22 3.25 0 7.84-.05 15.67-.05 23.51q0 24.94 0 49.9v1.44c-.11 1.73-.62 2.28-2.33 2.37-1.19.06-2.4 0-3.6 0-5.19 0-10.39 0-15.59 0-2.77 0-3 .28-3 3v18a18.86 18.86 0 0 1 -.43 4 5.84 5.84 0 0 1 -4.78 4.78 18.87 18.87 0 0 1 -4 .44q-22.79 0-45.58 0a16.69 16.69 0 0 1 -3.8-.4 6 6 0 0 1 -5.11-5.58 54.58 54.58 0 0 1 -.38-5.73q0-29.73-.03-59.48z"
-            fill="#70c599"
-          ></path>
-          <path
-            d="m326.16 390.85h-60.48c-1.28 0-2.56 0-3.83-.13-3.58-.35-5.45-2.13-6.13-5.66a17.86 17.86 0 0 1 -.28-3.34q0-26.52 0-53a18.73 18.73 0 0 1 .39-3.81 6.16 6.16 0 0 1 5.31-5.15 20.3 20.3 0 0 1 3.82-.3h122.86a20.34 20.34 0 0 1 3.58.28 6.61 6.61 0 0 1 5.63 5.19 14.84 14.84 0 0 1 .41 3.81v53.26a13.47 13.47 0 0 1 -.32 3.34c-.81 3.22-2.66 5-5.93 5.39a32.53 32.53 0 0 1 -3.83.15z"
-            fill="#2d3887"
-          ></path>
-          <path
-            d="m515.42 390.23q0-30.11 0-60.22c0-1.36 0-2.73.08-4.08.26-3.78 2.21-5.79 5.93-6.33a33.26 33.26 0 0 1 4.78-.3q21.24 0 42.47 0a30.29 30.29 0 0 1 4.77.38 5.94 5.94 0 0 1 5.39 5.27 28 28 0 0 1 .38 4.3q0 61.18 0 122.37a22 22 0 0 1 -.46 4.52 5.65 5.65 0 0 1 -4.56 4.69 15.71 15.71 0 0 1 -3.56.45q-23.5 0-47 0c-.48 0-1 0-1.44 0-4.3-.34-6.51-2.61-6.69-6.92-.06-1.36-.06-2.72-.06-4.08q-.03-30.04-.03-60.05z"
-            fill="#70c599"
-          ></path>
-          <path
-            d="m57.83 319.28c3.62 0 9.52.24 15.44.05 5.73-.19 9.73 2.67 12.23 7.67 3 5.85 6 11.66 9 17.49q14.91 29 29.86 57.94c1.73 3.32 3.22 6.57 3 10.42a15.51 15.51 0 0 1 -1.85 6.64c-6.57 12.06-12.68 24.37-18.93 36.61a10.06 10.06 0 0 1 -2 3c-3.81 3.53-9.42 2.51-11.9-2.22-2.7-5.17-5.34-10.37-8-15.56l-54.1-105.32a23.55 23.55 0 0 1 -2.46-5.69c-1.24-5.61 1.84-10.12 7.51-10.85a32.7 32.7 0 0 1 3.83-.15z"
-            fill="#a82175"
-          ></path>
-          <path
-            d="m327.37 399.43h32.38c2.72.07 3.25.56 3.26 3.24 0 4.48-.05 9-.05 13.44q0 17.75.05 35.5a23.35 23.35 0 0 1 -.15 3.35 6.62 6.62 0 0 1 -5.17 5.91 15.76 15.76 0 0 1 -4 .41h-53a14.44 14.44 0 0 1 -3.34-.26c-3.58-.86-5.51-3-5.89-6.63a21.31 21.31 0 0 1 -.08-2.4v-48.46c0-3.68.43-4.1 4.14-4.1z"
-            fill="#2d3887"
-          ></path>
-        </g>
-      </svg>
-      <h1>Urban Cities // Environment</h1>
-      <h2>8th - 9th February 2020</h2>
-      <h2>Singapore University of Technology and Design</h2>
-    </div>
-    <img class="hero-image" src="../assets/cityscape-light-theme.svg" alt="hero cityscape">
-    <svg
-      viewBox="0 0 363.71 144.51"
-      xmlns="http://www.w3.org/2000/svg"
-      class="cloud cloud-one"
-      width="160px"
-    >
-      <path
-        d="m-376.51 211.52q-84.71 0-169.42 0c-13 0-11.68 1.84-11.4-10.08.19-7.83 4.68-11.73 14-13.08 9-1.31 9.35-2 10.31-9.57 1.12-8.7.56-17.62 5.11-25.9 5.37-9.78 14.35-16 27-17.49 5.47-.64 11.52.71 17 2 9.15 2.08 9.36 2.19 9.37-5.39 0-15.61 8.53-28.58 18.81-40.91s24.86-20 42.45-22.77c18.62-3 35.65.83 50.15 11.29 16.16 11.66 26 26.66 32.59 43.2 3.28 8.17 2.92 8.19 11.32 3.42 11.58-6.58 24.59-6.73 37.49-4.41 24.18 4.35 38.41 17.94 46.62 36.46 1.12 2.54 1.31 5.38 1.76 8.1.57 3.38 2.68 4.92 6.93 5.19a74.72 74.72 0 0 1 13.21 2c11.94 3 17.65 10.28 18.77 20.17.55 4.78.36 9.62.62 14.44.12 2.32-1.22 3.24-3.83 3.31-3.33.09-6.66.32-10 .32q-84.45 0-168.91 0z"
-        fill="#fff"
-        transform="translate(557.49 -67.34)"
-      ></path>
-    </svg>
-    <svg
-      viewBox="0 0 344.31 163.86"
-      xmlns="http://www.w3.org/2000/svg"
-      class="cloud cloud-two"
-      width="60px"
-    >
-      <path
-        d="m-349.73 396.33c3.72-2.42 7-4.45 10.14-6.6 29.51-20.14 74.73-6.06 85.51 26.64.59 1.79 1.54 3.54 1.74 5.36.81 7.36 4.39 11.42 12.66 13.14 8.05 1.69 12.2 7.76 11.33 15.82-.59 5.44 1.55 8.73 6.51 11.72 9.25 5.57 11.27 16.32 5.19 24.74-.95 1.31-3.16 2.2-4.94 2.5a51.7 51.7 0 0 1 -8.55.36h-313c-14 0-15.15-1-14-14.33 1.27-15.37 4.1-30.24 15.29-42.63 5.92-6.55 12.57-11.05 22-10 7.6.85 9.56-2.77 10.81-9 3-14.91 5.33-30.17 10.64-44.39 16.33-43.72 55.86-53.71 98.94-33.41 22.71 10.69 33.91 30.22 44.11 50.53 1.09 2.18 2.13 4.38 3.27 6.54a29.69 29.69 0 0 0 2.35 3.01z"
-        fill="#fff"
-        transform="translate(557.49 -326.16)"
-      ></path>
-    </svg>
-    <svg
-      viewBox="0 0 210.22 92.74"
-      xmlns="http://www.w3.org/2000/svg"
-      class="cloud cloud-three"
-      width="120px"
-    >
-      <path
-        d="m-371.27 603.6c1.87 1 3.36 1.85 4.93 2.58 4.57 2.15 6 5.63 5.3 9.71-.16 1-1.81 2.17-3 2.48-1.68.43-3.63.11-5.46.11q-94.21 0-188.42.07c-4.76 0-7.74-1.24-9.93-4.89-8.26-13.73-.05-28 18-31 7.83-1.31 7.63-1.32 8.32-7.88.73-6.9 1.31-13.93 7.73-19.22 9.15-7.54 19.47-7.48 30.46-3.48 2.39.86 5.09 2.22 6.22-1.42 3.92-12.52 15.86-17.83 28.24-22.8 13.87-5.57 39.39.82 48.73 15.61a84.3 84.3 0 0 1 6.69 11.84c1.62 3.82 4 4.48 8.29 3.46 7.35-1.73 14.65 1.11 17.14 6.33.53 1.11 1 2.65.55 3.63-1.8 3.64-.44 4.53 3.93 4.85 17.23 1.27 24.15 18.23 14.49 26.58-1 .84-1.41 2.14-2.21 3.44z"
-        fill="#fff"
-        transform="translate(571.11 -525.86)"
-      ></path>
-    </svg>
-    <svg
-      viewBox="0 0 148.98 70.24"
-      xmlns="http://www.w3.org/2000/svg"
-      class="cloud cloud-four"
-      width="130px"
-    >
-      <path
-        d="m-460.3 814.78c-24 0-47.92-.06-71.88.06-4.21 0-6.62-1.21-7.13-5.11-.28-2.07-1.07-4.11-1.15-6.18-.49-11.7 6.4-16.65 19-13.74.65.15 1.26.41 1.9.57 6.14 1.51 6.69 1.09 7-4.78.65-11.7 4.15-22.46 13-31.4 14.56-14.7 42.16-11.27 51.64 6.42 3.5 6.55 3.41 6.18 11.61 4.46a31.41 31.41 0 0 1 12.66.23c8.05 1.74 10.84 7.77 12.3 14.41a23.89 23.89 0 0 1 .44 4c.24 5.22 1.67 9 8.47 10.58 7.22 1.62 10.28 7.25 10.9 13.7.5 5.3-1.5 7.15-7.3 7.16q-30.71 0-61.42 0z"
-        fill="#fff"
-        transform="translate(540.49 -744.98)"
-      ></path>
-    </svg>
-  </div>
-</template>
-
-<script>
-export default {
-  name: "Hero",
-  data() {
-    return {
-      theme: "light"
-    };
-  },
-  mounted: function() {
-    this.setInitialTheme();
-  },
-  methods: {
-    toggleTheme() {
-      if (this.theme === "dark") {
-        document.documentElement.setAttribute("theme", "light");
-        localStorage.setItem("theme", "light");
-        this.theme = "light";
-      } else {
-        document.documentElement.setAttribute("theme", "dark");
-        localStorage.setItem("theme", "dark");
-        this.theme = "dark";
-      }
-    },
-    setInitialTheme() {
-      if (localStorage.getItem("theme")) {
-        const cachedTheme = localStorage.getItem("theme");
-        document.documentElement.setAttribute("theme", cachedTheme);
-        this.theme = cachedTheme;
-      } else {
-        document.documentElement.setAttribute("theme", this.theme);
-      }
-    }
-  }
-};
-</script>
-
-<style scoped>
-#hero {
-  position: relative;
-  height: 100vh;
-  /* border: 5px solid blue; */
-}
-
-nav {
-  /* border: 2px solid red; */
-  padding: 25px 40px;
-}
-
-nav > ul {
-  display: flex;
-  align-items: center;
-}
-
-nav > ul > svg {
-  margin-right: auto;
-}
-
-nav > ul > svg > g {
-  fill: var(--color-text);
-  transition: fill 0.6s ease-out;
-}
-
-nav > ul > li + li {
-  margin-left: 30px;
-}
-
-nav a {
-  font-family: var(--font-primary), sans-serif;
-  text-decoration: none;
-  text-transform: uppercase;
-  color: var(--color-text);
-  transition: color 0.6s ease-out;
-}
-
-.nav-register-link {
-  border-bottom: 5px solid var(--color-text);
-  padding-bottom: 2px;
-  transform: translateY(4px);
-  transition: border-bottom 0.6s ease-out;
-}
-
-.hero-content {
-  /* border: 1px solid red; */
-  position: relative;
-  z-index: 1;
-}
-
-.hero-content > svg {
-  width: 54.5vw;
-  margin-left: 17vw;
-  margin-top: 4.5vw;
-  margin-bottom: 20px;
-}
-
-.hero-content > h1 {
-  font-family: proxima-soft;
-  font-size: 55px;
-  font-size: calc(3vw + 12px);
-  font-weight: 900;
-  color: var(--color-text);
-  margin-left: 34.75vw;
-  margin-bottom: 8px;
-  transition: color 0.6s ease-out;
-}
-
-.hero-content > h2 {
-  font-family: proxima-soft;
-  font-size: 24px;
-  font-size: calc(1vw + 10px);
-  font-weight: 700;
-  color: var(--color-text);
-  margin-left: 35vw;
-  transition: color 0.6s ease-out;
-}
-
-.hero-image {
-  position: absolute;
-  bottom: -1.5px;
-  opacity: 0.4;
-}
-
-.cloud {
-  position: absolute;
-  opacity: 0.5;
-}
-
-.cloud-one {
-  transform: translate(4vw, -5vw);
-  animation: cloud-animation-one 6s linear infinite alternate;
-}
-
-.cloud-two {
-  transform: translate(38vw, 1vw);
-  animation: cloud-animation-two 6s linear infinite alternate;
-}
-
-.cloud-three {
-  transform: translate(70vw, -2vw);
-  animation: cloud-animation-three 6s linear infinite alternate;
-}
-
-.cloud-four {
-  transform: translate(82vw, -80px);
-  animation: cloud-animation-four 6s linear infinite alternate;
-}
-
-@keyframes cloud-animation-one {
-  100% {
-    transform: translate(12vw, -5vw);
-  }
-}
-
-@keyframes cloud-animation-two {
-  100% {
-    transform: translate(28vw, 1vw);
-  }
-}
-
-@keyframes cloud-animation-three {
-  100% {
-    transform: translate(60vw, -2vw);
-  }
-}
-
-@keyframes cloud-animation-four {
-  100% {
-    transform: translate(88vw, -80px);
-  }
-}
-</style>
