+++ src/components/hero/HeroContent.vue
@@ -1,0 +1,79 @@
+<template>
+  <div class="hero-content">
+    <svg viewBox="0 0 551.39 142.09" xmlns="http://www.w3.org/2000/svg">
+      <g transform="translate(-27.85 -319.19)">
+        <path
+          d="m239 328.39a15.33 15.33 0 0 1 -2.06 7.06q-13 25.13-26 50.26-17.93 34.86-35.8 69.73a15.47 15.47 0 0 1 -2.06 3.22 7.07 7.07 0 0 1 -10.62.2 15.38 15.38 0 0 1 -2.46-3.8l-55.28-107.66q-3.61-7-7.25-14.06a10.23 10.23 0 0 1 -.73-8.56 7.54 7.54 0 0 1 5.84-5.11 15.48 15.48 0 0 1 3.33-.33c22.16 0 44.31-.07 66.46 0 10.4 0 20.79.24 31.18.23 8.08 0 16.15-.2 24.23-.34a16.09 16.09 0 0 1 4.74.52 8.52 8.52 0 0 1 6.48 8.64z"
+          fill="#a82175"
+        ></path>
+        <path
+          d="m418.7 390q0-29.76 0-59.5a43.67 43.67 0 0 1 .35-5.25 6.14 6.14 0 0 1 5.41-5.58 31.77 31.77 0 0 1 4.78-.38q21.58 0 43.18 0a28 28 0 0 1 4.76.41 6 6 0 0 1 5.15 4.72 14.47 14.47 0 0 1 .48 3.55v18.95c0 3.39.5 4 3.86 3.79 5.43-.37 10.86-.1 16.29-.21h1.2c2.72.06 3.22.55 3.22 3.25 0 7.84-.05 15.67-.05 23.51q0 24.94 0 49.9v1.44c-.11 1.73-.62 2.28-2.33 2.37-1.19.06-2.4 0-3.6 0-5.19 0-10.39 0-15.59 0-2.77 0-3 .28-3 3v18a18.86 18.86 0 0 1 -.43 4 5.84 5.84 0 0 1 -4.78 4.78 18.87 18.87 0 0 1 -4 .44q-22.79 0-45.58 0a16.69 16.69 0 0 1 -3.8-.4 6 6 0 0 1 -5.11-5.58 54.58 54.58 0 0 1 -.38-5.73q0-29.73-.03-59.48z"
+          fill="#70c599"
+        ></path>
+        <path
+          d="m326.16 390.85h-60.48c-1.28 0-2.56 0-3.83-.13-3.58-.35-5.45-2.13-6.13-5.66a17.86 17.86 0 0 1 -.28-3.34q0-26.52 0-53a18.73 18.73 0 0 1 .39-3.81 6.16 6.16 0 0 1 5.31-5.15 20.3 20.3 0 0 1 3.82-.3h122.86a20.34 20.34 0 0 1 3.58.28 6.61 6.61 0 0 1 5.63 5.19 14.84 14.84 0 0 1 .41 3.81v53.26a13.47 13.47 0 0 1 -.32 3.34c-.81 3.22-2.66 5-5.93 5.39a32.53 32.53 0 0 1 -3.83.15z"
+          fill="#2d3887"
+        ></path>
+        <path
+          d="m515.42 390.23q0-30.11 0-60.22c0-1.36 0-2.73.08-4.08.26-3.78 2.21-5.79 5.93-6.33a33.26 33.26 0 0 1 4.78-.3q21.24 0 42.47 0a30.29 30.29 0 0 1 4.77.38 5.94 5.94 0 0 1 5.39 5.27 28 28 0 0 1 .38 4.3q0 61.18 0 122.37a22 22 0 0 1 -.46 4.52 5.65 5.65 0 0 1 -4.56 4.69 15.71 15.71 0 0 1 -3.56.45q-23.5 0-47 0c-.48 0-1 0-1.44 0-4.3-.34-6.51-2.61-6.69-6.92-.06-1.36-.06-2.72-.06-4.08q-.03-30.04-.03-60.05z"
+          fill="#70c599"
+        ></path>
+        <path
+          d="m57.83 319.28c3.62 0 9.52.24 15.44.05 5.73-.19 9.73 2.67 12.23 7.67 3 5.85 6 11.66 9 17.49q14.91 29 29.86 57.94c1.73 3.32 3.22 6.57 3 10.42a15.51 15.51 0 0 1 -1.85 6.64c-6.57 12.06-12.68 24.37-18.93 36.61a10.06 10.06 0 0 1 -2 3c-3.81 3.53-9.42 2.51-11.9-2.22-2.7-5.17-5.34-10.37-8-15.56l-54.1-105.32a23.55 23.55 0 0 1 -2.46-5.69c-1.24-5.61 1.84-10.12 7.51-10.85a32.7 32.7 0 0 1 3.83-.15z"
+          fill="#a82175"
+        ></path>
+        <path
+          d="m327.37 399.43h32.38c2.72.07 3.25.56 3.26 3.24 0 4.48-.05 9-.05 13.44q0 17.75.05 35.5a23.35 23.35 0 0 1 -.15 3.35 6.62 6.62 0 0 1 -5.17 5.91 15.76 15.76 0 0 1 -4 .41h-53a14.44 14.44 0 0 1 -3.34-.26c-3.58-.86-5.51-3-5.89-6.63a21.31 21.31 0 0 1 -.08-2.4v-48.46c0-3.68.43-4.1 4.14-4.1z"
+          fill="#2d3887"
+        ></path>
+      </g>
+    </svg>
+    <h1>Urban Cities // Environment</h1>
+    <h2>8th - 9th February 2020</h2>
+    <h2>Singapore University of Technology and Design</h2>
+  </div>
+</template>
+
+<script>
+export default {
+  name: "hero-content"
+};
+</script>
+
+<style scoped>
+.hero-content {
+  /* border: 1px solid red; */
+  position: relative;
+  z-index: 1;
+}
+
+.hero-content > svg {
+  width: 54.5vw;
+  margin-left: 17vw;
+  margin-top: 4.5vw;
+  margin-bottom: 20px;
+}
+
+.hero-content > h1 {
+  font-family: proxima-soft;
+  font-size: 55px;
+  font-size: calc(3vw + 12px);
+  font-weight: 900;
+  color: var(--color-text);
+  margin-left: 34.75vw;
+  margin-bottom: 8px;
+  transition: color 0.6s ease-out;
+}
+
+.hero-content > h2 {
+  font-family: proxima-soft;
+  font-size: 24px;
+  font-size: calc(1vw + 10px);
+  font-weight: 700;
+  color: var(--color-text);
+  margin-left: 35vw;
+  transition: color 0.6s ease-out;
+}
+</style>
+
+
