+++ src/components/hero/HeroImage.vue
@@ -1,0 +1,118 @@
+<template>
+  <div>
+    <img class="hero-image" src="../../assets/cityscape-light-theme.svg" alt="hero cityscape">
+    <svg
+      viewBox="0 0 363.71 144.51"
+      xmlns="http://www.w3.org/2000/svg"
+      class="cloud cloud-one"
+      width="160px"
+    >
+      <path
+        d="m-376.51 211.52q-84.71 0-169.42 0c-13 0-11.68 1.84-11.4-10.08.19-7.83 4.68-11.73 14-13.08 9-1.31 9.35-2 10.31-9.57 1.12-8.7.56-17.62 5.11-25.9 5.37-9.78 14.35-16 27-17.49 5.47-.64 11.52.71 17 2 9.15 2.08 9.36 2.19 9.37-5.39 0-15.61 8.53-28.58 18.81-40.91s24.86-20 42.45-22.77c18.62-3 35.65.83 50.15 11.29 16.16 11.66 26 26.66 32.59 43.2 3.28 8.17 2.92 8.19 11.32 3.42 11.58-6.58 24.59-6.73 37.49-4.41 24.18 4.35 38.41 17.94 46.62 36.46 1.12 2.54 1.31 5.38 1.76 8.1.57 3.38 2.68 4.92 6.93 5.19a74.72 74.72 0 0 1 13.21 2c11.94 3 17.65 10.28 18.77 20.17.55 4.78.36 9.62.62 14.44.12 2.32-1.22 3.24-3.83 3.31-3.33.09-6.66.32-10 .32q-84.45 0-168.91 0z"
+        fill="#fff"
+        transform="translate(557.49 -67.34)"
+      ></path>
+    </svg>
+    <svg
+      viewBox="0 0 344.31 163.86"
+      xmlns="http://www.w3.org/2000/svg"
+      class="cloud cloud-two"
+      width="60px"
+    >
+      <path
+        d="m-349.73 396.33c3.72-2.42 7-4.45 10.14-6.6 29.51-20.14 74.73-6.06 85.51 26.64.59 1.79 1.54 3.54 1.74 5.36.81 7.36 4.39 11.42 12.66 13.14 8.05 1.69 12.2 7.76 11.33 15.82-.59 5.44 1.55 8.73 6.51 11.72 9.25 5.57 11.27 16.32 5.19 24.74-.95 1.31-3.16 2.2-4.94 2.5a51.7 51.7 0 0 1 -8.55.36h-313c-14 0-15.15-1-14-14.33 1.27-15.37 4.1-30.24 15.29-42.63 5.92-6.55 12.57-11.05 22-10 7.6.85 9.56-2.77 10.81-9 3-14.91 5.33-30.17 10.64-44.39 16.33-43.72 55.86-53.71 98.94-33.41 22.71 10.69 33.91 30.22 44.11 50.53 1.09 2.18 2.13 4.38 3.27 6.54a29.69 29.69 0 0 0 2.35 3.01z"
+        fill="#fff"
+        transform="translate(557.49 -326.16)"
+      ></path>
+    </svg>
+    <svg
+      viewBox="0 0 210.22 92.74"
+      xmlns="http://www.w3.org/2000/svg"
+      class="cloud cloud-three"
+      width="120px"
+    >
+      <path
+        d="m-371.27 603.6c1.87 1 3.36 1.85 4.93 2.58 4.57 2.15 6 5.63 5.3 9.71-.16 1-1.81 2.17-3 2.48-1.68.43-3.63.11-5.46.11q-94.21 0-188.42.07c-4.76 0-7.74-1.24-9.93-4.89-8.26-13.73-.05-28 18-31 7.83-1.31 7.63-1.32 8.32-7.88.73-6.9 1.31-13.93 7.73-19.22 9.15-7.54 19.47-7.48 30.46-3.48 2.39.86 5.09 2.22 6.22-1.42 3.92-12.52 15.86-17.83 28.24-22.8 13.87-5.57 39.39.82 48.73 15.61a84.3 84.3 0 0 1 6.69 11.84c1.62 3.82 4 4.48 8.29 3.46 7.35-1.73 14.65 1.11 17.14 6.33.53 1.11 1 2.65.55 3.63-1.8 3.64-.44 4.53 3.93 4.85 17.23 1.27 24.15 18.23 14.49 26.58-1 .84-1.41 2.14-2.21 3.44z"
+        fill="#fff"
+        transform="translate(571.11 -525.86)"
+      ></path>
+    </svg>
+    <svg
+      viewBox="0 0 148.98 70.24"
+      xmlns="http://www.w3.org/2000/svg"
+      class="cloud cloud-four"
+      width="130px"
+    >
+      <path
+        d="m-460.3 814.78c-24 0-47.92-.06-71.88.06-4.21 0-6.62-1.21-7.13-5.11-.28-2.07-1.07-4.11-1.15-6.18-.49-11.7 6.4-16.65 19-13.74.65.15 1.26.41 1.9.57 6.14 1.51 6.69 1.09 7-4.78.65-11.7 4.15-22.46 13-31.4 14.56-14.7 42.16-11.27 51.64 6.42 3.5 6.55 3.41 6.18 11.61 4.46a31.41 31.41 0 0 1 12.66.23c8.05 1.74 10.84 7.77 12.3 14.41a23.89 23.89 0 0 1 .44 4c.24 5.22 1.67 9 8.47 10.58 7.22 1.62 10.28 7.25 10.9 13.7.5 5.3-1.5 7.15-7.3 7.16q-30.71 0-61.42 0z"
+        fill="#fff"
+        transform="translate(540.49 -744.98)"
+      ></path>
+    </svg>
+  </div>
+</template>
+
+<script>
+export default {
+  name: "hero-image"
+};
+</script>
+
+<style scoped>
+.hero-image {
+  position: absolute;
+  bottom: -1.5px;
+  opacity: 0.4;
+}
+
+.cloud {
+  position: absolute;
+  opacity: 0.5;
+}
+
+.cloud-one {
+  transform: translate(4vw, -5vw);
+  animation: cloud-animation-one 6s linear infinite alternate;
+}
+
+.cloud-two {
+  transform: translate(38vw, 1vw);
+  animation: cloud-animation-two 6s linear infinite alternate;
+}
+
+.cloud-three {
+  transform: translate(70vw, -2vw);
+  animation: cloud-animation-three 6s linear infinite alternate;
+}
+
+.cloud-four {
+  transform: translate(82vw, -80px);
+  animation: cloud-animation-four 6s linear infinite alternate;
+}
+
+@keyframes cloud-animation-one {
+  100% {
+    transform: translate(12vw, -5vw);
+  }
+}
+
+@keyframes cloud-animation-two {
+  100% {
+    transform: translate(28vw, 1vw);
+  }
+}
+
+@keyframes cloud-animation-three {
+  100% {
+    transform: translate(60vw, -2vw);
+  }
+}
+
+@keyframes cloud-animation-four {
+  100% {
+    transform: translate(88vw, -80px);
+  }
+}
+</style>
+
+
