+++ src/components/hero/Nav.vue
@@ -1,0 +1,165 @@
+<template>
+  <nav>
+    <ul>
+      <svg viewBox="0 0 551.39 142.09" xmlns="http://www.w3.org/2000/svg" width="150px">
+        <g fill="#2d3887" transform="translate(-27.85 -319.19)">
+          <path
+            d="m239 328.39a15.33 15.33 0 0 1 -2.06 7.06q-13 25.13-26 50.26-17.93 34.86-35.8 69.73a15.47 15.47 0 0 1 -2.06 3.22 7.07 7.07 0 0 1 -10.62.2 15.38 15.38 0 0 1 -2.46-3.8l-55.28-107.66q-3.61-7-7.25-14.06a10.23 10.23 0 0 1 -.73-8.56 7.54 7.54 0 0 1 5.84-5.11 15.48 15.48 0 0 1 3.33-.33c22.16 0 44.31-.07 66.46 0 10.4 0 20.79.24 31.18.23 8.08 0 16.15-.2 24.23-.34a16.09 16.09 0 0 1 4.74.52 8.52 8.52 0 0 1 6.48 8.64z"
+          ></path>
+          <path
+            d="m418.7 390q0-29.76 0-59.5a43.67 43.67 0 0 1 .35-5.25 6.14 6.14 0 0 1 5.41-5.58 31.77 31.77 0 0 1 4.78-.38q21.58 0 43.18 0a28 28 0 0 1 4.76.41 6 6 0 0 1 5.15 4.72 14.47 14.47 0 0 1 .48 3.55v18.95c0 3.39.5 4 3.86 3.79 5.43-.37 10.86-.1 16.29-.21h1.2c2.72.06 3.22.55 3.22 3.25 0 7.84-.05 15.67-.05 23.51q0 24.94 0 49.9v1.44c-.11 1.73-.62 2.28-2.33 2.37-1.19.06-2.4 0-3.6 0-5.19 0-10.39 0-15.59 0-2.77 0-3 .28-3 3v18a18.86 18.86 0 0 1 -.43 4 5.84 5.84 0 0 1 -4.78 4.78 18.87 18.87 0 0 1 -4 .44q-22.79 0-45.58 0a16.69 16.69 0 0 1 -3.8-.4 6 6 0 0 1 -5.11-5.58 54.58 54.58 0 0 1 -.38-5.73q0-29.73-.03-59.48z"
+          ></path>
+          <path
+            d="m326.16 390.85h-60.48c-1.28 0-2.56 0-3.83-.13-3.58-.35-5.45-2.13-6.13-5.66a17.86 17.86 0 0 1 -.28-3.34q0-26.52 0-53a18.73 18.73 0 0 1 .39-3.81 6.16 6.16 0 0 1 5.31-5.15 20.3 20.3 0 0 1 3.82-.3h122.86a20.34 20.34 0 0 1 3.58.28 6.61 6.61 0 0 1 5.63 5.19 14.84 14.84 0 0 1 .41 3.81v53.26a13.47 13.47 0 0 1 -.32 3.34c-.81 3.22-2.66 5-5.93 5.39a32.53 32.53 0 0 1 -3.83.15z"
+          ></path>
+          <path
+            d="m515.42 390.23q0-30.11 0-60.22c0-1.36 0-2.73.08-4.08.26-3.78 2.21-5.79 5.93-6.33a33.26 33.26 0 0 1 4.78-.3q21.24 0 42.47 0a30.29 30.29 0 0 1 4.77.38 5.94 5.94 0 0 1 5.39 5.27 28 28 0 0 1 .38 4.3q0 61.18 0 122.37a22 22 0 0 1 -.46 4.52 5.65 5.65 0 0 1 -4.56 4.69 15.71 15.71 0 0 1 -3.56.45q-23.5 0-47 0c-.48 0-1 0-1.44 0-4.3-.34-6.51-2.61-6.69-6.92-.06-1.36-.06-2.72-.06-4.08q-.03-30.04-.03-60.05z"
+          ></path>
+          <path
+            d="m57.83 319.28c3.62 0 9.52.24 15.44.05 5.73-.19 9.73 2.67 12.23 7.67 3 5.85 6 11.66 9 17.49q14.91 29 29.86 57.94c1.73 3.32 3.22 6.57 3 10.42a15.51 15.51 0 0 1 -1.85 6.64c-6.57 12.06-12.68 24.37-18.93 36.61a10.06 10.06 0 0 1 -2 3c-3.81 3.53-9.42 2.51-11.9-2.22-2.7-5.17-5.34-10.37-8-15.56l-54.1-105.32a23.55 23.55 0 0 1 -2.46-5.69c-1.24-5.61 1.84-10.12 7.51-10.85a32.7 32.7 0 0 1 3.83-.15z"
+          ></path>
+          <path
+            d="m327.37 399.43h32.38c2.72.07 3.25.56 3.26 3.24 0 4.48-.05 9-.05 13.44q0 17.75.05 35.5a23.35 23.35 0 0 1 -.15 3.35 6.62 6.62 0 0 1 -5.17 5.91 15.76 15.76 0 0 1 -4 .41h-53a14.44 14.44 0 0 1 -3.34-.26c-3.58-.86-5.51-3-5.89-6.63a21.31 21.31 0 0 1 -.08-2.4v-48.46c0-3.68.43-4.1 4.14-4.1z"
+          ></path>
+        </g>
+      </svg>
+      <li>
+        <a href="#">About</a>
+      </li>
+      <li>
+        <a href="#">Prizes</a>
+      </li>
+      <li>
+        <a href="#">Schedule</a>
+      </li>
+      <li>
+        <a href="#">Sponsors</a>
+      </li>
+      <li>
+        <a href="#">FAQ</a>
+      </li>
+      <li class="nav-register-link">
+        <a href="#">Register</a>
+      </li>
+      <li
+        style="height: 40px; border: 1px solid var(--color-text); transition: border 0.6s ease-out;"
+      ></li>
+      <li>
+        <svg
+          v-if="theme === 'light'"
+          width="25px"
+          style="transform: translateY(3.5px); cursor: pointer;"
+          @click="toggleTheme"
+          viewBox="0 0 222.59 218.11"
+          xmlns="http://www.w3.org/2000/svg"
+        >
+          <g transform="translate(-195.2 -286.74)">
+            <path
+              d="m376.78 472.87c.41 10.77.42 10.74-9.93 8.59-5.68-1.17-11.41-2.15-17.08-3.39-2.91-.63-4.47.29-5.81 3-3.94 7.94-8.21 15.73-12.45 23.78-7.29-5.36-14.5-10.5-21.51-15.91-2.48-1.91-4.22-2.22-6.83-.12-6 4.83-12.41 9.16-18.43 14-2.76 2.2-3.79 1.52-5.19-1.34-3.27-6.72-7.07-13.18-10.28-19.92-1.48-3.11-3.19-4.13-6.54-3.4-7.13 1.57-14.39 2.58-21.46 4.36-4.19 1.06-5.52-.06-5.13-4.14.17-1.81.07-3.65.08-5.47.76-3.9.15-7.83.35-11.74 0-2.44-.06-4.9.05-7.31-.41 1.05.41 2.44-.7 3.38 0-4.17-2.25-5.7-6.2-6.53-6.95-1.45-13.69-3.83-20.62-5.39-3.31-.75-3.58-1.89-2.28-4.7 3.33-7.2 6.33-14.56 9.67-21.76 1.17-2.5.85-3.59-1.37-5.62-6.63-6.05-14.48-10.81-19.92-18.16 6.67-5.85 13.16-11.92 20.11-17.42 2.32-1.84 2.29-3 1.27-5.25-3.16-6.93-6-14-9.36-20.85-1.95-4-1-5.35 3.15-6.24 6.66-1.42 13.13-3.76 19.81-5.07 4-.77 5.49-2.78 5.74-6.41 1.17.93.14 2.57.75 3.42-.23-2.3 0-4.78-.1-7.25-.2-4 .41-7.92-.35-11.84 0-1-.13-2-.11-3 .12-8.09.1-7.95 8.08-6.08 6.43 1.5 13 2.58 19.45 3.91 2.44.51 4-.19 5.14-2.53 3.63-7.26 7.56-14.38 11.15-21.67 1.24-2.51 2.29-2.58 4.26-.87 6.4 5.55 13.33 10.38 20.2 15.32 1.94 1.4 2.61.92 3.91 0 6.7-5 13.47-9.92 20.08-15 2.37-1.84 3.48-2 4.91 1.09 2.86 6.13 6.25 12 9.4 18s6.16 6.91 12.65 4.95c6.15-1.85 12.57-2.8 18.88-4.1 1.59-.33 2.56.27 2.54 2.07v8c-.76 3.9-.15 7.83-.35 11.74v7c.36-.9-.46-2.29.65-3.23-.3 6 4.55 6.16 8.37 7.22 6.51 1.82 13.08 3.38 19.64 5 3.11.79 1.61 2.83 1 4.28-2.62 6.26-5.4 12.45-8.13 18.65s-3.25 6.85 1.7 10.84c6.15 5 11.75 10.56 18.13 15.23-5.42 7.39-13.3 12.13-20 18.19-2.31 2.09-2.42 3.21-1.28 5.65 3.22 6.91 6.11 14 9.32 20.89 1.56 3.36 1 4.94-2.85 5.79-6.63 1.47-13.07 3.75-19.71 5.11-3.95.81-6.25 2.35-6.23 6.53-1.17-.93-.14-2.57-.75-3.42.23 2.3.05 4.78.1 7.25.26 4.02-.37 8.02.41 11.89z"
+              fill="#f4c425"
+            ></path>
+            <path d="m236.22 318.13c1.74 3.87.42 7.95.82 11.91-.32.47-.64.45-1 0z" fill="#f3c001"></path>
+            <path
+              d="m376 330c.41-4-.92-8 .82-11.91q.06 6 .14 11.91c-.36.51-.68.49-.96 0z"
+              fill="#f3c001"
+            ></path>
+            <path
+              d="m236.08 330h1v8h-.71c-.13-1.39-.27-2.79-.41-4.18.04-1.27.04-2.52.12-3.82z"
+              fill="#f4c214"
+            ></path>
+            <path
+              d="m376 330h1c.05 1.26.11 2.51.16 3.76-.14 1.39-.28 2.79-.41 4.18h-.75z"
+              fill="#f4c214"
+            ></path>
+            <path
+              d="m237 461c-.41 4 .92 8-.82 11.91q-.06-6-.14-11.91c.36-.49.68-.51.96 0z"
+              fill="#f3c001"
+            ></path>
+            <path
+              d="m376.78 472.87c-1.78-3.87-.42-7.95-.78-11.87.32-.45.64-.47 1 0q-.15 5.92-.22 11.87z"
+              fill="#f3c001"
+            ></path>
+            <path
+              d="m237 461h-1c-.05-1.26-.11-2.51-.16-3.76.14-1.39.28-2.79.41-4.18h.75z"
+              fill="#f4c214"
+            ></path>
+            <path
+              d="m376.92 461h-1v-8h.71c.13 1.39.27 2.79.41 4.18-.04 1.27-.04 2.52-.12 3.82z"
+              fill="#f4c214"
+            ></path>
+            <path
+              d="m232.09 387a75.38 75.38 0 0 1 64.91-65.91 76 76 0 0 0 18.94 0 75.38 75.38 0 0 1 64.97 65.91 60.84 60.84 0 0 0 0 16.94 75.11 75.11 0 0 1 -65.91 65.97 60.84 60.84 0 0 0 -16.94 0 75.1 75.1 0 0 1 -65.97-65.91 60.84 60.84 0 0 0 0-17z"
+              fill="#f7a705"
+            ></path>
+            <g fill="#f7a500">
+              <path d="m316 321.09a39.15 39.15 0 0 1 -18.94 0z"></path>
+              <path d="m232.09 387a31.24 31.24 0 0 1 0 16.94z"></path>
+              <path d="m380.91 404a31.24 31.24 0 0 1 0-16.94z"></path>
+              <path d="m298 469.91a31.24 31.24 0 0 1 16.94 0z"></path>
+            </g>
+          </g>
+        </svg>
+        <svg
+          v-else
+          width="25px"
+          style="transform: translateY(3.5px); cursor: pointer;"
+          @click="toggleTheme"
+          viewBox="0 0 223.99 225"
+          xmlns="http://www.w3.org/2000/svg"
+        >
+          <path
+            d="m106 0h13l19.6 2.67a9.73 9.73 0 0 1 -4.24 2.55c-39.09 14.31-62.72 54.67-55 93.86 12.47 62.94 86 90.76 137.25 51.92 2.36-1.78 4.89-3.34 7.34-5v1c-.93 1.87-2 3.7-2.78 5.63-13.6 33.75-38.06 55.91-72.49 66.88-8.92 2.85-18.47 3.71-27.73 5.49h-12.95a22.42 22.42 0 0 0 -3.33-.8c-47.58-4.11-86.17-35-100-80.4-2.16-7.08-3.14-14.52-4.67-21.8v-19c.39-2.44.7-4.89 1.17-7.32 8.61-44.68 34.38-74.68 77.25-89.59 8.82-3.09 18.36-4.09 27.58-6.09z"
+            fill="#f3f3fb"
+            transform="translate(-.01)"
+          ></path>
+        </svg>
+      </li>
+    </ul>
+  </nav>
+</template>
+
+<script>
+export default {
+  name: "nav"
+};
+</script>
+
+<style scoped>
+nav {
+  /* border: 2px solid red; */
+  padding: 25px 40px;
+}
+
+nav > ul {
+  display: flex;
+  align-items: center;
+}
+
+nav > ul > svg {
+  margin-right: auto;
+}
+
+nav > ul > svg > g {
+  fill: var(--color-text);
+  transition: fill 0.6s ease-out;
+}
+
+nav > ul > li + li {
+  margin-left: 30px;
+}
+
+nav a {
+  font-family: var(--font-primary), sans-serif;
+  text-decoration: none;
+  text-transform: uppercase;
+  color: var(--color-text);
+  transition: color 0.6s ease-out;
+}
+
+.nav-register-link {
+  border-bottom: 5px solid var(--color-text);
+  padding-bottom: 2px;
+  transform: translateY(4px);
+  transition: border-bottom 0.6s ease-out;
+}
+</style>
